Fix Spec-transformation for Dredd
The Standard-OpenApi3-Specs have to be manipulated before taken to Dredd because of some incompatibilities in Dredd. One was that Dredd does not support multiple examples. But we want them, so we split the Spec file into multiple ones before running Dredd.

In this process there was a bug: If a endpoint had one field with only one example and another with multiple examples, it didn't appeared in the result file.