Fix a bug in the login process of group-monitor.
During the login of a group-monitor, all members of the group get created in the db if not already existent. The Bug did refresh their tokens if they were already present. The consequence was, that when a member logged in before the group monitor, there session was destroyed, and they have log in again.