Add backup environment file to gitignore
The e2e tests use a special environment file for the frontend. The 
existing one is copied before running the tests and copied back 
afterwards. In case the tests fail the copy-back step is not done and 
the file still exists which may cause confusion. Git ignoring it should 
lessen the confusion.