Fix absurd memory usage while validating files.
Because every resource was loaded into the memory, a validation could take a lot of memory which resulted in failing uploads on large workspaces.

This was easily fixable by only opening them when the content is used. In case of validation it is used, but only for players, not for resources (and those are the files which can get very big).