Further refactoring of the relationship between workspace and DB.
Until now there were a complete separation of everything wich has to do with files and everything which has to do with db and onl the controllers my use classes of both worlds an merge data. This is not longer possible since the WorkspaceValidator validates against data from the db as well because of because of performance (and logic-) problems.

On the plus side, we can get rid of TesttakersFolder class.

Also: Fix some bugs which where revealed by fixing the tests.