Flash broadcastingService (delete internal cash and reconnect) when Testtakers.XMLs are uploaded or deleted.
This seems not to be a very good user-experience at fist sight because every active group-monitor and testee will have to reconnect when a Testtakers.XML will be uploaded or deleted. Shouldn't exact those TestSessions whose Logins changed / where deleted in the XML be updated in BS's cache?
Yes, but it would be *extremly* complicated (and bug-affine) to catch all the possible types of changing like: adding/removes codes or bookeltes, deleting or restoring logins at all, change groups etc. pp. It would even be impossible to detect  tests which belong to deleted logins in the db without changing the structure and storing the login-name again in person_sessions or login_sessions.
While we don't need all this complexity in polling-mode it's better just to flash everything when logins get changed, even it results to some reconnections and be safe that all monitors will show what's really going on. The testcenter can automatically reconnect so, there is no problem.