Use 'slim' base images for docker containers
For frontend we use a self-built base container which has some
dependencies already installed. This container is located at DockerHUB
for now to make it available outside the HU-network.
The Dockerfile for this base image is added to the repo under /docker.
This way it can also be built locally if pulling from either dockerhub
or gitlab is not possible.