Rework docker containers for development
With development containers the node_modules folder is now built at
compile time (Dockerfile) and saved in a temporary folder (app-temp).
At runtime (i.e. compose up) the contents of the temp folder are copied
over to the bind mount folder so they are availbale to the host system
as well. This makes working with IDE possible.#

Also:
- Don't ignore node_modules folder for docker. The folder needs to be
mounted into containers.
- File permissions are fixed by running 'make fix-docker-user'
- Add lock files for all package.jsons