[integration-tests] Stabilize tests
When doing re-login there was a problem: The admin-starter tried to load the workspace-data, but when logout was already clicked, the call returned a 401 and the error-box blocked the login button. So the test failed whenever running fast enough. So we have to assert the existence oif the workspace button before continuing. For some reason it was impossible to use tehe selector "cy.contains('sample_workspace')" for some reason - it always existed somewhere in the dom. That's why we use the data-cy attribute at this point as well.