Rework system tests
Now running in an extra container instead of within the frontend. 
Therefore also has it's own directory etc.
The container is part of the usual dev setup but only run on command 
(changed the Makefile target).
It is necessary to change the backend address used by the frontend when 
running system tests within the docker network. Therefore the 
environment file is changed before the tests and changed back after. 
This process may be refined later.

Things that are no longer working but may be re-introduced later are: 
coverage reports, running cypress visually within a container and having 
different modes for the used backend data when running system tests.