Revert system test setup to old state
where it is contained in it's own compose setup with a separate DB and 
non-volume data directory. This way no existing test data can be 
overwritten or deleted by the system tests.

A lot of tests still fail, but since some still work fine it is assumed 
that this is a matter of bad test design and selectors and not the 
container setup.