[Frontend] Reset app error-state after closing error box.
Otherwise the last error (after a canceled reload for example) gets replayed under some circumstances when TestControllerComponent gets reloaded.

The whole architecture with the error-observable is questionable though and will be redesigned soon. For now this important fix is enough to prohibit a lot of misbehaviour when changing users and tests fast.