[Frontend] Remove bogus and complete useless localStorage-functionality.
The AuthData has to be stored in localStorage, but it was questioned from there every single time it was used! Under some race conditions this could lead to errors, and we got the wrong authData.

The idea to store the TestConfig in localStorage was abandoned long time ago (if ever implemented correctly) and there were only functionless ruins left which are now removed.