[Backend] Add _PIPES_AS_CONCAT_ to sql-mode.
This is necessary because we can not use the (non sql-standard) substitution _concat_ for _||_ for string concatenation, because all queries have to work in both, SQLite and MySQL.

It's a good idea to define the sql-mode explicitly anyway. Wehn they change the defaults this might break some queries as making   _ONLY_FULL_GROUP_BY_ actually did in the past.