[Backend] Fix the check for duplicate logins.
Now the first one counts (like in older versions). We made it regard both files as invalid, when they contain the same login, because we didn't want to decide which one is correct. Since it is not possible to create this with new uploads meanwhile, we don't have to take care so much. The first one counts.