[Backend] Fix Dredd-tests
Due to an unknown reason the system throws a segmentation fault, when an endpoint is targeted with POST and Content-Type:multipart/form-data but without body. This is not good, but could not be debugged even with gdb. So it's most likely a bug in php or one of its extensions.
It's not a very likely scenario anyway, but it occurred in the API-tests because there were no content attached when testing the error cases. This could be fixed by hooking into the error cases and attach some upload file as we do it in the success cases. It's a more realistic test this way anyway.