Add container-scanning to CI pipeline
- Only triggered by creating prod build (tags)
- Split up build jobs for the 3 components
This is necessary because the scanning report can only be a single file. 
By having different jobs, the 3 reports are automatically merged (and 
shown on the "Security" tab in GitLab).
- For scanning the tool "trivy" is used.
"exit-code 0" means that the job never fails even though issues might 
have been found. This is necessary because GitLab does not accept 
artifacts (reports) from failed jobs.
The format paramaters are for GitLab, so it can show the results on the 
"Security" tab of the pipeline.