[Backend] Prevent SQL-patches to fail silently
Remove the commit/rollback logic because create table, modify columns etc. - the main content of sql patch files - does implicit commits anyway and this logic causes more confusion than it helps. It's a shame that a error-throwing statement in a patch can leave a database in between two patches. But at least now, we would get correctly informed about that.