Rename config/cert_config.yml to ssl-config.yml
The file now contains more SSL settings than just the certificates. 
Therefore it's name is generalized.

The new settings are a minimum TLS version of 1.2 and specific Cypher 
Suites which are deemed secure.

Correspondingly adjust the installer and updater. The file is no longer 
generated by the installer but downloaded as all other files.

In the compose config this file is explicitly used instead of a whole 
directory. This should be safer and less error-prone.