Infrastructure: Update npm packages for runner
Result of running 'npm audit fix'. The packages decode-uri-component and 
json5 were having issues and are updated.

Other critical issues are still present but can't be fixed since the 
packages are old and unmaintained. This is the case for gulp and dredd.
This is not much of an issue, since these packages are only used for 
development.