[Backend] prevent duplicate login_sessions
There can be multiple person_sessions attached to one login_session, but it should be only one login_session by name. When two people logged in at the very same moment with the same login it could be created twice. This update prevents this by making the "name"-column unique and exchange the insert and get statements. So now the new session gets inserted (if it is not there), and then, when it was there, it gets retrieved.