[Setup] Remove CORS and Cache headers from backend config
Allowing CORS is not necessary as the backend is only used by our 
frontend, living on the same server and in some cases by 
server-to-server communication, which is not affected by those headers 
anyway.

It is not clear why the Caching was disabled in the first place. It is 
useful and enabled again.