be: Re-enable caching of api-calls for calls that return 410.
Browser-site caching of api-calls was removed because it seemed to be useless, but now there is at least one scenario where it causes trouble: when a login is expired, and then was prolonged again (common scenario during development). Since the browsers does store the 410-call, the account would not be accessible ver again without clearing the browser's cache.