fe: Fix malfunctioning test-loading after being logged in as workspace-admin.
In the workspace-admin a setting on the router was made, that prevented the loading of a test without full page reload. The setting was made to force a reload of the files (or results) list in the workspace admin when ws number in url is changed manually but had the critical side effect to sabotage the test-loader somehow. So it's removed and the reloading of the file-list is is now solved with an observable.