fe: Don't show an error message when test-loading-precess was aborted
by clicking the logo for example. Errors could occur on various places when the process was aborted because roottestlet was already removed but it still tried to load some units or resources. Now we throw a standardized error if rootTestlet is missing at any point, and ignore this type of error if it occurs while loading.

This leaved open a very theoretical scenario where rootTestlet is missing during loading-process but for another reason than abortion: in this case the error might be swallowed. But it's very unlikely because how should it take place? if there was a problem with the booklet it would alleady have bailed out.