be: Allow retry when creating a person-session with an already present suffix.
It's extremely unlikely (1 to 1971788797440) in production, but in test context when randomizer is static, it occurs by definition. Since we safe costs and use not-crypto-safe random here, it might be more likely in prod (eg if two persons should be created in the very same nanosecond) so it's not bad to have a fallback here.