xsd: Allow arbitrary string for `id` in `<Variable>`.
Resolves https://github.com/iqb-berlin/testcenter/issues/339

We don't take the Regex from the JSON-Schema-based definition of the variable-list, because the current Teststudio allowed arbitrary strings as ids. There are already files around with ids containing a hyphen, umlauts, dots and whatnot so for now we allow everything.

The XML-based variable-list will be deprecated soon in favor of the XML-schema based one.