[docker] Remove healthchecks again
After further investigation, health checks were found to be unsuitable for our purposes, as they continue to run indefinitely even when the status is "healthy". This can be mitigated by using the "start_time" and "start_interval" parameters, but these are not yet implemented in the current compose.

There are currently two use cases for health checks:
a) In the CI, the database was sometimes so slow that even waiting for the backend was not long enough. This health check was retained but moved to docker-compose.dev.
b) Waiting for all services before starting e2e tests: these health checks were removed again, and the old wait-on-server script was reinstated.