FE: Testcontroller: Get rid of redundant array unitDefinitionTypes.
The unitDefinitionType was initially introduced as a separate concept to the player-id, because it seemed to be possible, that there a different player (or at least versions) which can play the same types of unit definitions. In this sense there is a unitDefinitionType field in the vopStartCommand in Verona. Later we simplified it, so that playerId is always the same that unitDefinitionType.