be: Use unique-constraint to ensure uniqueness of logins instead of xs:ID.
xs:ID as a datatype does have some constraints which apparently where a problem for some users: you where not allowed to start the name with a number.
So we use a generic string now and ensure the uniqueness with a <xs:unique>-constraint which is more correct anyway.

This had interesting consequences: Since a XML with duplicate logins gets recognized as invalid XML anyway, we can remove our own check for duplicate logins.