[fe] Fix page navigation panel in syscheck and refactor page nav
Use the same paging component in the syscheck and the normal unit host.

Change the way page data is handled: The Verona API serves pages as a 
dictionary with page number/indexes as key and page labels as value. 
This is now changed to a simple array of labels - the papge index is 
implicit. At some point the API will change this but for now it is 
transformed in the frontend code.

- Refactor page variables: knownPages is now pageList with simplified 
type
- Some small UI and code formatting improvements