fe: Fix navigation when there are no inro- or outro-units.
There was an (informal) rule to have at least on unit at the start and at the end of a booklet which does not belong to a testlet. The navigation behaviour in others cases was undefined: We should we navigate when everything is locked by time for example. This fix lets the TC navigate to the locked unit anyway if there is no alternative, to display the correct status (like "locked by time").

fixes #480