Add technology for additional constraints in XML-formats
which unfortunately can not be modelled with XMLSchema 1.0.

Example: Every state Option and corresponding state that is referred in a <Show>-Restriction must be defined in the <States>-object in Booklet.xml.

This could be easily tested with the <xsd:assert>-element in XMLSchema 1.1.

But XMLSchema 1.1 is not supported by PHP (nor by IDEA, although there is a setting for it), and there is no plugin or suitable external library for this:
 * the saxon/c open source version does not allow XMLschema 1.1 either
 * xerces only have bindings for java and perl.