Remove confusing <DefaultOption>-tag from xsd.
This was a bad idea, because the default option is the first empty `<Option>` (`DefaultOption`)-tag, and if there was more than one the Tag could be confusing. The only reason it was introduces was that xs:assert was missing and there were no way checking for at least one empty `<Option>`-tag with XMLschema 1.0. But now we have a replacement for that.