If in test-mode display it in the logs and return it to frontend to display it in error messages as well
When CyPress-tests fail headless locally or on CI, it fails sometimes because of a /state, /log or similar call wich gets rejected by the backend. One suspicion is that - due to a bug in CypPress itself - the TestMode don't get attached, and thus the main db is used not the test db. To see if this happened the error mode is visible in any error message now.