Include TestMode into the disconnectNotificationUri of subscribed endpoints
When a websocket connection to the broadcasting-service is canceled, the latter records this to the backend (This is the only case where the stream of information is reversed and the broadcasting service calls the backend instead the other way round). When doing system/e2e tests we had the problem that this internal call from bs to be didn't contain the TestMode header, so the backend looked up in the wrong DB for the canceled test. And thus the logs where full of 'could not send connection-lost signal to...'