Frontend: use unitId as SessionId in communication between player and host.
Reason: Newer versions of the Aspect-player sends a last `vopStateChangedNotification` on `window:unload`. In this (and possible other, undiscovered) situations, a stateUpdate might come after a new `vopStartCommand` and the data gets attached to the wring unit!

Solution: Use the unitId (precisely the thing called unitAlias with is the truly unique identifier of the unit) as sessionId, so every vopStateChangedNotification can te glued to the correct unit.

This is the intentional usage of the sessionId. Quote from the Verona Specification:
> The session id helps to link the correct unit with the data of the message.