[e2e] Improve stability of test-controller: nav-restriction-testlet
Set TestIsolation true and remove unnecessary afterEach routines. Its faster and cleaner

Add a parent describe: Testing the presentation/response-complete test can be unstable. If you do not want to use this test, you can disable it with "skip".

Sometimes the test headless failed, because presentation-/response-complete is not there yet. For this reason, the waiting time has been extended