[e2e] Improve stability of session-management: hot-restart-mode
Remove unnecessary waiting times for responses-complete, it is not necessary to give a response for a session entry in the response file.

Use an interception: Make sure the session has been saved.

Set test-isolation true and remove logouts

Interception for response download for more stability

Make the test code clearer