[be] Enforce correct json_encoded string
When persisting $newState in the db, the resulting string should always represent a JSON object, not falsely a '[]'.

The json_encode flag JSON_FORCE_OBJECT is too blunt, as this would also enforce objects within the $newState entries. The content is dynamic and not knowable by us. So we stay conservative and only enforce the first layer of php array with (object) casting.

zh #1125