[fe] Implement header component and service
This new component is for displaying a unified header, so that the
header is not re-implemented in every component.

The service is for allowing components to set values for the header from
everywhere in the app.

The switch to using this new component is not finished, This is just the
foundation to be built upon.