[fe] Rework global page layouting
Global layout is now done with flex instead of absolute positioning.

CSS classes for page header and body have been removed and is now done
with a small class in the root component.

Several components which used this global styling are changed to work
with the new styling model.

The group monitor needed the most changes, but should work largely the
same. It is also using flex layout in some places, but not all. Only the
connection indicator had to be moved, since the component no longer
draws its own header. It is moved into another toolbar. The position
stays roughly the same.

The global CSS file for layouting the monitor layout has been moved to
the attachment manager as it is only used there. Styling for the group
monitor has been moved to the components CSS, where it belongs.