[fe] Implement Material theming infrastructure
Material theming now works by having a theme that uses a generated color
palette as base. Having a base color palette is mandatory.

Base colors are then overridden via SCSS variables
"theme-color-overrides.scss"

The colors to use when overriding are defined in "theme-colors.scss".
There are 3 variants defined which can be swtched by changing the
"current-theme" variable. Later this will be possible at runtime, but
for now it only works at compile time.

The last file is specifically to override component styles. It is by now
only done for the header background color.