[fe] Rework review UI
Now uses a sidenav instead of a dialog. This way it is easier to see the
unit content while writing.

Also restructured and restyled the input fields. The keyboard detection
stuff has been removed as the textarea can have the same size regardless
of target device.

The test-controller component now has 2 alternative sidenavs (unit list
and review panel). If both are defined separately and positioned at
"end" the Material backdrop break. Therefore another switch is added to
show one or the other.

#649