[be, fe, docs] Add 'download review' button to starter component
Only shows up if the "review" claim is part of the session and therefore a 'run-review' login type.
Adds 'review' as new claim type.
hook.js needs to be updated to add the CSV Byte Order Mark (BOM) to the expected output
Rename accessObjects to claims in starter component to match data structure