[e2e] Fix backdrop click on unit menu
The unit menu supposedly now overlays the backdrop, when Cypress wants
to click it to close the sidenav. The sidenav has been changed to
contain either the unit menu or the review menu. Somehow this has caused
this. Manual testing shows clearly that the backdrop works fine. There
is no better solution than to force right now.