[be] Remove check on weather tests and units exist for getReviews to work
This check fails if a reviewer looks at the booklets and tries to show all past reviews. This would have thrown an error, because the 'run-review' mode itself does not update state and therefore doesn't create rows in the 'tests' or 'units' table by design.

For getting reviews we do not check for written 'units' or 'tests' rows now.